--- src/hooks/api/tag/useTagQuery.ts
@@ -1,61 +1,0 @@
-import { useMutation, useQuery, useQueryClient } from 'react-query';
-
-import { del, get, post } from '~/libs/api/client';
-
-interface TagResponseInterface {
-  message: string;
-  data: TagDataResponseInterface;
-}
-
-interface TagDataResponseInterface extends PaginationInterface {
-  content: TagInterface[];
-}
-
-// TODO: constants로 묶어서 관리?
-const TAG_LIST_QUERY_KEY = 'tags';
-
-interface UseTagQueryProps {
-  pageNumber: number;
-}
-
-export default function useTagQuery({ pageNumber }: UseTagQueryProps) {
-  const queryClient = useQueryClient();
-
-  const refresh = () => {
-    queryClient.invalidateQueries([TAG_LIST_QUERY_KEY]);
-  };
-
-  // NOTE: useInfiniteQuery 사용?
-  const fetchTags = (pageNumber: number = 1) =>
-    get<TagResponseInterface>(
-      `/v1/tag/list?paged=true&offset=20&pageSize=20&pageNumber=${pageNumber}`
-    );
-
-  const query = useQuery<TagResponseInterface>(
-    [TAG_LIST_QUERY_KEY, pageNumber],
-    () => fetchTags(pageNumber),
-    {
-      keepPreviousData: true,
-    }
-  );
-
-  const createTagMutation = useMutation((content: string) => post('/v1/tag/add', { content }), {
-    onSuccess: () => {
-      refresh();
-    },
-  });
-
-  const deleteTagMutation = useMutation((id: number) => del(`/v1/tag/remove/${id}`), {
-    onSuccess: () => {
-      refresh();
-    },
-  });
-
-  return {
-    tags: query.data ? query.data.data.content : [],
-    isLoading: query.isLoading,
-    query,
-    createTag: createTagMutation.mutate,
-    deleteTag: deleteTagMutation.mutate,
-  };
-}
